consolidate redactApiKeys to utils
