+++ apps/sim/stores/panel/console/store.ts
@@ -1,44 +1,11 @@
 import { create } from 'zustand'
 import { devtools, persist } from 'zustand/middleware'
+import { redactApiKeys } from '@/lib/utils'
 import { useChatStore } from '../chat/store'
 import { ConsoleEntry, ConsoleStore } from './types'
 
 // MAX across all workflows
 const MAX_ENTRIES = 50
-
-/**
- * Recursively redacts API keys in an object
- * @param obj The object to redact API keys from
- * @returns A new object with API keys redacted
- */
-const redactApiKeys = (obj: any): any => {
-  if (!obj || typeof obj !== 'object') {
-    return obj
-  }
-
-  if (Array.isArray(obj)) {
-    return obj.map(redactApiKeys)
-  }
-
-  const result: Record<string, any> = {}
-
-  for (const [key, value] of Object.entries(obj)) {
-    // Check if the key is 'apiKey' (case insensitive)
-    if (
-      key.toLowerCase() === 'apikey' ||
-      key.toLowerCase() === 'api_key' ||
-      key.toLowerCase() === 'access_token'
-    ) {
-      result[key] = '***REDACTED***'
-    } else if (typeof value === 'object' && value !== null) {
-      result[key] = redactApiKeys(value)
-    } else {
-      result[key] = value
-    }
-  }
-
-  return result
-}
 
 /**
  * Gets a nested property value from an object using a path string
